Add new way of handling secrets via new replacement pattern
Secrets can be declared in a secrets section in the root of a
fabfile (similar to a question) and be used in any host-configuration
as `%secret.MY_SECRET_NAME%. Phab will inject the actual value when
the host configuration is requested. The actual value can be passed
via environment variable of command line argument `--secret key=value`
As a last ressort, the user is asked for the secret via shell.