fix(k8s): the real bug which prevents scaffolding k8s files
Versions prior 3.7.9 worked correctly, they scaffolded the files into
the right folder, relative to the fabfile. 3.7.10 fixed a slightly
related bug that previous scaffolded files were not removed before
running the scaffolding again which made it impossible to remove k8s
definition files.

But this revealed a very profound bug in K8sMethod which used a wrong
folder for scaffolding where the user had no access to. This is now
fixed by using a correct root-path for the scaffolder.