fix(script-execution-context) Create unique project-name for docker-compose-run
there might be situations where docker-compose's name-spacing is not
enough to separate multiple runs from each other. To reduce conflicts
when a script is run via the `docker-compose-run`-context, multiple
times on the same host, phab creates a unique project name for each run
based on radnomness to reduce possible namespace-collisions. docker-
compose is just using the name of the current folder to namspace an
application, that might not be enough in ci scenarios, where the same
folder can be checked out to a completely different path. These changes
should mitigate the chance of a collision here.